Fix for issue 57 (Timezone should allow for fractional offsets) in trunk

+++ RestFB/source/library/com/restfb/types/User.java
@@ -73,7 +73,7 @@ public class User extends NamedFacebookType {
   private String picture;
 
   @Facebook
-  private Integer timezone;
+  private Double timezone;
 
   @Facebook
   private Boolean verified;
@@ -405,7 +405,7 @@ public class User extends NamedFacebookType {
    * 
    * @return The user's timezone offset.
    */
-  public Integer getTimezone() {
+  public Double getTimezone() {
     return timezone;
   }
